Add callerId to allow outgoing phone calls.
* First issue: I hadn't set the App URL up correctly on Twilio,
forgetting the /voice endpoint. This resulted in a 405 error, method
(POST) not allowed. That was a little confusing.

* Second issue: That fixed, Twilio would show on the Console > Debugger
that there was no caller ID when originating the call from the browser.
This was fixed by including a verified Twilio phone number from the
backend through to the frontent, and passing it along to the Twilio app
as a parameter (app.js).